This isn't working... there must be another way
I’m thinking that there must be another way to do this. I can’t bind
the events for the dynamic regions… I’ve tried it many different ways.
Maybe the answer is to use a collection view with layout views as
children. The issue I have there is that collection views require
collections… I guess my collection could be “computer users” or
something like that. I was trying to implement this functionality
anyway. This was one of my original ideas, but I decided against it for
some reason. Now I think it’s the only way.